refactor: 기존 ReservationService를 도메인별로 분리
